refactor(ui): split cockpit.rs into panels, overlays and theme modules
cockpit.rs had grown to ~3260 lines. Move-only split, no behaviour
change:

- ui/cockpit.rs keeps the render() entry point, the 4-panel layout
  and the status bar (~190 lines).
- ui/panels/{probe,inventory,mannies,scanner}.rs — one file per
  panel, each with its height helper; scanner keeps its detail-line
  builders (sector_object_lines, sector_interest_color).
- ui/overlays/ — one file per wizard overlay; pickers.rs groups the
  seven pick-list-based ones; mod.rs hosts the shared centered_rect
  and render_pick_list helpers and re-exports the render functions.
- ui/theme.rs — colours, icons, labels and the format_duration /
  format_age helpers.

Functions crossing module boundaries become pub(crate); the public
surface of ui:: is unchanged (cockpit::render).
