+++ src/ui/mod.rs
@@ -1,1 +1,4 @@
 pub mod cockpit;
+pub(crate) mod overlays;
+pub(crate) mod panels;
+pub(crate) mod theme;
