+++ src/ui/panels/mod.rs
@@ -1,0 +1,9 @@
+pub(crate) mod inventory;
+pub(crate) mod mannies;
+pub(crate) mod probe;
+pub(crate) mod scanner;
+
+pub(crate) use inventory::{inventory_panel_height, render_inventory_panel};
+pub(crate) use mannies::render_mannies_panel;
+pub(crate) use probe::{probe_panel_height, render_probe_panel};
+pub(crate) use scanner::render_scanner_panel;
